Added 404 response to repo endpoint

+++ app/controllers/index_controller.py
@@ -15,12 +15,16 @@ def index():
 @index_blueprint.route("/<repository_name>")
 def view(repository_name):
     repository = Repository.find(current_app.config.get("REPOS_DIR_PATH"), repository_name)
-    for change in repository.added_files():
-        print(change)
+    if repository:
+        for change in repository.added_files():
+            print(change)
 
-    changes = {}
-    changes["added_documents"] = [change.a_path for change in repository.added_files()]
-    changes["deleted_documents"] = [change.a_path for change in repository.deleted_files()]
-    changes["modified_documents"] = [change.a_path for change in repository.modified_files()]
-    return Response(json.dumps(changes, sort_keys=True), mimetype="application/json")
+        changes = {}
+        changes["added_documents"] = [change.a_path for change in repository.added_files()]
+        changes["deleted_documents"] = [change.a_path for change in repository.deleted_files()]
+        changes["modified_documents"] = [change.a_path for change in repository.modified_files()]
+        return Response(json.dumps(changes, sort_keys=True), mimetype="application/json")
+    else:
+        error = { "error": "Not found!" }
+        return Response(json.dumps(error, sort_keys=True), mimetype="application/json"), 404
